+++ Website/webb/src/common/components/Navbars/NavBar.module.css
@@ -6,7 +6,6 @@
   text-align: center;
   align-items: center;
   margin-bottom: 0px;
-  border: 2px solid #2397e9;
   border-radius: 0%;
   display: flex;
   flex-direction: column;
@@ -39,8 +38,4 @@
   align-items: center;
   justify-content: flex-start;
   flex-direction: column;
-}
-
-button {
-  color: red;
 }
